Move getCurrentUserTraits selector to its own file
