+++ packages/mobile/src/analytics/saga.ts
@@ -1,29 +1,6 @@
 import { call, select, spawn, take } from 'redux-saga/effects'
-import { createSelector } from 'reselect'
-import { nameSelector } from 'src/account/selectors'
+import { getCurrentUserTraits } from 'src/analytics/selectors'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
-import { userLocationDataSelector } from 'src/networkInfo/selectors'
-import { defaultCurrencySelector } from 'src/stableToken/selectors'
-import { accountAddressSelector, currentAccountSelector } from 'src/web3/selectors'
-
-export const getCurrentUserTraits = createSelector(
-  [
-    currentAccountSelector,
-    accountAddressSelector,
-    defaultCurrencySelector,
-    nameSelector,
-    userLocationDataSelector,
-  ],
-  (walletAddress, accountAddress, currency, name, { countryCodeAlpha2 }) => {
-    return {
-      accountAddress,
-      walletAddress,
-      currency,
-      name,
-      countryCodeAlpha2,
-    }
-  }
-)
 
 export function* updateUserTraits() {
   let prevTraits
